+++ packages/lumx-core/src/js/components/Icon/constants.ts
@@ -1,0 +1,3 @@
+import { Size } from '@lumx/core/js/constants';
+
+export const ICON_SIZES = [Size.xxs, Size.xs, Size.s, Size.m, Size.l, Size.xl, Size.xxl];